qos: Speed up plist traversal in pm_qos_set_value_for_cpus()
The plist is already sorted and traversed in ascending order of PM QoS
value, so we can simply look at the lowest PM QoS values which affect
the given request's CPUs until we've looked at all of them, at which
point the traversal can be stopped early. This also lets us get rid of
the pesky qos_val array.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: //. <174224977+skye-tachyon@users.noreply.github.com>